# apps/api/src/routes/v1/company.ts
import type { FastifyInstance, FastifyRequest } from 'fastify';

import type { CompanyActivityPageOptions, CompanyProjectionSources } from '@agentos/database';
import type { CompanyProjectionPort, PlatformProvidersPort } from '../../gateway/ports.js';
import type { CredentialStore } from '../../gateway/principal.js';
import { authenticate, requireOperator } from '../../gateway/principal.js';
import type { GatewayRuntime } from '../../gateway/ports.js';
import { correlationIdOf, fail, replyFailure } from '../../gateway/http.js';
import type { DemoReadinessPort } from './demo-readiness.js';
import type { ProvisioningRoutePort } from './provisioning.js';
import { mapAiTeam } from '../../projections/ai-team.js';
import { mapAttention } from '../../projections/attention.js';
import { mapActivity } from '../../projections/activity.js';
import { mapOverview, type WorkspaceOverviewSource } from '../../projections/overview.js';
import {
  companyActivityRouteSchema,
  companyAiTeamRouteSchema,
  companyAttentionRouteSchema,
  companyOverviewRouteSchema,
  registerOpenApiSchemas,
} from './openapi-schemas.js';

export interface CompanyRouteDependencies {
  readonly runtime: GatewayRuntime;
  readonly credentials: CredentialStore;
  readonly projections: CompanyProjectionPort;
  readonly readiness?: DemoReadinessPort;
  readonly provisioning?: ProvisioningRoutePort;
  readonly providers?: PlatformProvidersPort;
  readonly enabledModules?: readonly string[];
}

function queryLimit(value: unknown): number | undefined {
  if (value === undefined) return undefined;
  const parsed = Number(value);
  if (!Number.isSafeInteger(parsed) || parsed < 1 || parsed > 200) {
    fail('VALIDATION_FAILED', 'limit must be an integer between 1 and 200');
  }
  return parsed;
}


type CompanyRouteSources = CompanyProjectionSources & {
  readonly provider?: { readonly configured: boolean; readonly provider: string | null; readonly demo?: boolean };
  readonly readiness_connectors?: readonly { readonly key: string; readonly class: string }[];
  readonly enabled_modules?: readonly string[];
  readonly module_capabilities?: readonly { readonly capability_id: string; readonly status: string }[];
  readonly workspace?: WorkspaceOverviewSource;
};

interface CompanyReadinessSource {
  readonly provider: { readonly configured: boolean; readonly provider: string | null; readonly demo?: boolean };
  readonly readiness_connectors: readonly { readonly key: string; readonly class: string }[];
}

function capabilityRows(value: unknown): readonly { readonly capability_id: string; readonly status: string }[] | undefined {
  if (!Array.isArray(value)) return undefined;
  const rows: { capability_id: string; status: string }[] = [];
  for (const item of value) {
    if (typeof item !== 'object' || item === null || Array.isArray(item)) continue;
    const record = item as Record<string, unknown>;
    if (typeof record['capability_id'] !== 'string' || typeof record['status'] !== 'string') continue;
    rows.push({ capability_id: record['capability_id'], status: record['status'] });
  }
  return rows;
}

async function readinessSource(
  dependencies: CompanyRouteDependencies,
  tenant_id: string,
): Promise<CompanyReadinessSource | undefined> {
  const [snapshot, providerRows] = await Promise.all([
    dependencies.readiness === undefined
      ? Promise.resolve(null)
      : dependencies.readiness.snapshot({ tenant_id }),
    dependencies.providers === undefined ? Promise.resolve([]) : dependencies.providers.list(),
  ]);
  const configuredProvider = providerRows[0];
  if (snapshot === null && configuredProvider === undefined) return undefined;
  const providerName = configuredProvider?.provider ?? snapshot?.provider.provider ?? null;
  return {
    provider: {
      configured: configuredProvider?.configured ?? snapshot?.provider.configured ?? false,
      provider: providerName,
      ...(configuredProvider?.mode === 'DEMO_MOCK' ? { demo: true } : {}),
    },
    readiness_connectors: snapshot === null
      ? []
      : [
          { key: 'erp', class: snapshot.connectors.erp.class },
          { key: 'events', class: snapshot.connectors.events.class },
        ],
  };
}

async function sourceWithShell(
  dependencies: CompanyRouteDependencies,
  tenant_id: string,
  enrich = true,
  activityPage?: CompanyActivityPageOptions,
): Promise<CompanyRouteSources> {
  const sources = activityPage === undefined
    ? await dependencies.projections.getSources(tenant_id)
    : await dependencies.projections.getSources(tenant_id, activityPage);
  if (!enrich) return sources;
  const [readiness, shell] = await Promise.all([
    readinessSource(dependencies, tenant_id),
    dependencies.provisioning === undefined ? Promise.resolve(null) : dependencies.provisioning.getShell(tenant_id),
  ]);
  const ownerById: Record<string, { readonly input_id: string; readonly status: 'UNRESOLVED' }> = {};
  for (const owner of sources.owner_inputs) {
    if (owner.status === 'UNRESOLVED') ownerById[owner.input_id] = { input_id: owner.input_id, status: 'UNRESOLVED' };
  }
  for (const input_id of shell?.unresolved_owner_inputs ?? []) {
    if (ownerById[input_id] === undefined) ownerById[input_id] = { input_id, status: 'UNRESOLVED' };
  }
  const capabilities = capabilityRows(shell?.capabilities);
  const erpClass = readiness?.readiness_connectors.find((connector) => connector.key === 'erp')?.class;
  const provisionedConnector = sources.connectors.some((connector) => /bound|connected|live|active/i.test(connector.status));
  const workspace: WorkspaceOverviewSource = {
    status: shell?.status ?? 'UNKNOWN',
    profile_ready: shell !== null,
    erp_ready: erpClass !== undefined && !/unconfig|missing|not_configured|unbound|absent/i.test(erpClass),
    knowledge_ready: (capabilities ?? []).some((capability) => /knowledge/i.test(capability.capability_id) && capability.status === 'ENABLED'),
    llm_ready: readiness?.provider.configured ?? false,
    channels_ready: provisionedConnector,
    test_agents_ready: sources.agents.length > 0,
  };
  return {
    ...sources,
    ...(readiness === undefined ? {} : { provider: readiness.provider, readiness_connectors: readiness.readiness_connectors }),
    ...(dependencies.enabledModules === undefined ? {} : { enabled_modules: dependencies.enabledModules }),
    ...(capabilities === undefined ? {} : { module_capabilities: capabilities }),
    ...(shell === null && readiness === undefined ? {} : { workspace }),
    owner_inputs: Object.values(ownerById),
  };
}

async function audit(
  request: FastifyRequest,
  runtime: GatewayRuntime,
  tenant_id: string,
  operation: string,
  detail: Record<string, unknown>,
  principal_kind: string,
  operator_id: string | undefined,
): Promise<void> {
  await runtime.audit.record({
    tenant_id,
    correlation_id: correlationIdOf(request, runtime),
    operation,
    principal_kind,
    ...(operator_id === undefined ? {} : { operator_id }),
    outcome: 'ACCEPTED',
    detail,
  });
}

/** Registers the five tenant company-console projections. */
export function registerCompanyRoutes(app: FastifyInstance, dependencies: CompanyRouteDependencies): void {
  registerOpenApiSchemas(app);
  const preHandler = authenticate(dependencies);
  app.get('/company/attention', { preHandler, schema: companyAttentionRouteSchema }, async (request, reply) => {
    const runtime = dependencies.runtime;
    try {
      const principal = requireOperator(request, 'telemetry:read');
      const sources = await sourceWithShell(dependencies, principal.tenant_id);
      const items = mapAttention(sources);
      await audit(request, runtime, principal.tenant_id, 'company.attention', { count: items.length }, principal.kind, principal.operator_id);
      return reply.code(200).send({ items });
    } catch (error) {
      return replyFailure(reply, error, correlationIdOf(request, runtime));
    }
  });

  app.get('/company/ai-team', { preHandler, schema: companyAiTeamRouteSchema }, async (request, reply) => {
    const runtime = dependencies.runtime;
    try {
      const principal = requireOperator(request, 'telemetry:read');
      const sources = await sourceWithShell(dependencies, principal.tenant_id);
      const agents = mapAiTeam(sources);
      await audit(request, runtime, principal.tenant_id, 'company.ai-team', { count: agents.length }, principal.kind, principal.operator_id);
      return reply.code(200).send({ agents });
    } catch (error) {
      return replyFailure(reply, error, correlationIdOf(request, runtime));
    }
  });

  app.get<{ Querystring: { limit?: string; cursor?: string } }>('/company/activity', { preHandler, schema: companyActivityRouteSchema }, async (request, reply) => {
    const runtime = dependencies.runtime;
    try {
      const principal = requireOperator(request, 'run:read');
      const limit = queryLimit(request.query.limit);
      const activityPage: CompanyActivityPageOptions = {
        ...(limit === undefined ? {} : { limit }),
        ...(request.query.cursor === undefined ? {} : { cursor: request.query.cursor }),
      };
      const sources = await sourceWithShell(dependencies, principal.tenant_id, false, activityPage);
      const activity = mapActivity(sources, activityPage);
      await audit(request, runtime, principal.tenant_id, 'company.activity', { count: activity.items.length }, principal.kind, principal.operator_id);
      return reply.code(200).send(activity);
    } catch (error) {
      return replyFailure(reply, error, correlationIdOf(request, runtime));
    }
  });


  app.get('/company/overview', { preHandler, schema: companyOverviewRouteSchema }, async (request, reply) => {
    const runtime = dependencies.runtime;
    try {
      const principal = requireOperator(request, 'telemetry:read');
      const sources = await sourceWithShell(dependencies, principal.tenant_id);
      const overview = mapOverview(sources);
      await audit(request, runtime, principal.tenant_id, 'company.overview', {
        attention_count: overview.attention.length,
        today_metric_count: overview.today?.metrics.length ?? 0,
        activity_count: overview.activity.length,
        workspace_checklist: overview.workspace?.checklist.length ?? 0,
      }, principal.kind, principal.operator_id);
      return reply.code(200).send(overview);
    } catch (error) {
      return replyFailure(reply, error, correlationIdOf(request, runtime));
    }
  });
}