# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentOS API",
    "description": "AgentOS gateway API contract.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "CompanyAttentionItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "severity",
          "domain",
          "title_key",
          "params",
          "href",
          "source_ref"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "danger"
            ]
          },
          "domain": {
            "type": "string",
            "enum": [
              "marketing",
              "sales",
              "care",
              "platform"
            ]
          },
          "title_key": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "href": {
            "type": "string"
          },
          "source_ref": {
            "type": "string"
          }
        }
      },
      "CompanyAttentionGroup": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "severity",
          "domain",
          "count",
          "title_key",
          "params",
          "href",
          "cta_key"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "danger"
            ]
          },
          "domain": {
            "type": "string",
            "enum": [
              "marketing",
              "sales",
              "care",
              "platform"
            ]
          },
          "count": {
            "type": "integer",
            "minimum": 1
          },
          "title_key": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "href": {
            "type": "string"
          },
          "cta_key": {
            "type": "string"
          }
        }
      },
      "CompanyAiTeamAgent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "domain",
          "status",
          "enabled",
          "readiness"
        ],
        "properties": {
          "domain": {
            "type": "string",
            "enum": [
              "marketing",
              "sales",
              "care"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PAUSED",
              "DISABLED",
              "NOT_READY",
              "NO_DATA"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "readiness": {
            "type": "string",
            "enum": [
              "READY",
              "NOT_READY",
              "UNKNOWN"
            ]
          },
          "runs_today": {
            "type": "integer"
          },
          "pending_approvals": {
            "type": "integer"
          },
          "open_handoffs": {
            "type": "integer"
          }
        }
      },
      "CompanyAiTeamStripEntry": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "domain",
          "status",
          "reason_key",
          "href"
        ],
        "properties": {
          "domain": {
            "type": "string",
            "enum": [
              "marketing",
              "sales",
              "care"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PAUSED",
              "DISABLED",
              "NOT_READY",
              "NO_DATA"
            ]
          },
          "reason_key": {
            "type": "string"
          },
          "counter_key": {
            "type": "string",
            "enum": [
              "runs_today",
              "pending_approvals",
              "open_handoffs"
            ]
          },
          "counter_value": {
            "type": "integer",
            "minimum": 1
          },
          "href": {
            "type": "string"
          }
        }
      },
...
      "CompanyOverviewToday": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "updated_at",
          "metrics"
        ],
        "properties": {
          "updated_at": {
            "type": "string"
          },
          "metrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyTodayMetric"
            }
          }
        }
      },
      "CompanyWorkspaceChecklistItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "key",
          "label_key",
          "done",
          "href"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "label_key": {
            "type": "string"
          },
          "done": {
            "type": "boolean"
          },
          "href": {
            "type": "string"
          }
        }
      },
      "CompanyOverviewWorkspace": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status",
          "checklist"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "checklist": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyWorkspaceChecklistItem"
            }
          }
        }
      },
      "CompanyOverviewSections": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "attention",
          "ai_team",
          "today",
          "activity",
          "workspace"
        ],
        "properties": {
          "attention": {
            "type": "string",
            "enum": [
              "OK",
              "ERROR"
            ]
          },
          "ai_team": {
            "type": "string",
            "enum": [
              "OK",
              "ERROR"
            ]
          },
          "today": {
            "type": "string",
            "enum": [
              "OK",
              "ERROR"
            ]
          },
          "activity": {
            "type": "string",
            "enum": [
              "OK",
              "ERROR"
            ]
          },
          "workspace": {
            "type": "string",
            "enum": [
              "OK",
              "ERROR"
            ]
          }
        }
      },
...
      "CompanyActivityItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "kind",
          "sentence_key",
          "params",
          "run_id",
          "domain",
          "occurred_at"
        ],
        "properties": {
          "kind": {
            "type": "string"
          },
          "sentence_key": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "run_id": {
            "type": "string"
          },
          "domain": {
            "type": "string",
            "enum": [
              "marketing",
              "sales",
              "care",
              "platform"
            ]
          },
          "occurred_at": {
            "type": "string"
          }
        }
      },
...
      "CompanyAttentionResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyAttentionItem"
            }
          }
        }
      },
      "CompanyAiTeamResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "agents"
        ],
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyAiTeamAgent"
            }
          }
        }
      },
      "CompanyActivityResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "next_cursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyActivityItem"
            }
          },
          "next_cursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CompanyOverviewResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "attention",
          "ai_team",
          "activity",
          "sections"
        ],
        "properties": {
          "attention": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyAttentionGroup"
            }
          },
          "ai_team": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyAiTeamStripEntry"
            }
          },
          "today": {
            "$ref": "#/components/schemas/CompanyOverviewToday"
          },
          "activity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyActivityItem"
            }
          },
          "workspace": {
            "$ref": "#/components/schemas/CompanyOverviewWorkspace"
          },
          "sections": {
            "$ref": "#/components/schemas/CompanyOverviewSections"
          }
        }
      },
...
      "PlatformSkillCatalogItem": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "skill_id",
          "runs_24h",
          "success_rate_24h",
          "p95_ms_24h"
        ],
        "properties": {
          "skill_id": {
            "type": "string"
          },
          "runs_24h": {
            "type": "integer",
            "minimum": 0
          },
          "success_rate_24h": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              {
                "type": "null"
              }
            ]
          },
          "p95_ms_24h": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "PlatformSkillCatalogResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "catalog"
        ],
        "properties": {
          "catalog": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformSkillCatalogItem"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error_code",
          "message",
          "retryable",
          "correlation_id"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "correlation_id": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/platform/skill-catalog": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSkillCatalogResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform/skill-catalog/{id}/entitlement/{tenant_id}": {
      "put": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tenant_id",
            "required": true
          }
        ],
        "responses": {
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/company/attention": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyAttentionResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/company/ai-team": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyAiTeamResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/company/activity": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyActivityResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/company/overview": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOverviewResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}